add utils file for util funcs
