+++ 2021-08-11-2021-08-11-logs.json
@@ -1,0 +1,1 @@
+[{"_id":"6113a5c929435d44355d191f","accountId":"6106c9362e4b3e7f21f0b113","jobId":"6113a4cd29435dde745d18f5","slot":"steps.Initializing Process.logs","payload":"Running Codefresh YAML validation\r\n","time":1628677577565},{"_id":"6113a5c929435d4d825d1920","accountId":"6106c9362e4b3e7f21f0b113","jobId":"6113a4cd29435dde745d18f5","slot":"steps.Initializing Process.logs","payload":"YAML validation passed successfully\r\n","time":1628677577625},{"_id":"6113a65f29435d0bfc5d192e","accountId":"6106c9362e4b3e7f21f0b113","jobId":"6113a4cd29435dde745d18f5","slot":"steps.Initializing Process.logs","payload":"YAML validation passed successfully\r\n","time":1628677727919},{"_id":"6113a65f29435dc9935d192d","accountId":"6106c9362e4b3e7f21f0b113","jobId":"6113a4cd29435dde745d18f5","slot":"steps.Initializing Process.logs","payload":"Running Codefresh YAML validation\r\n","time":1628677727805}]
