+++ 2021-08-11-2021-08-11-metadata.json
@@ -1,0 +1,1 @@
+[{"_id":"6113a5c9ec84f6519302852d","accountId":"6106c9362e4b3e7f21f0b113","jobId":"6113a4cd29435dde745d18f5","steps":{"Initializing Process":{"status":"pending","finishTimeStamp":"","creationTimeStamp":""}},"lastUpdate":1628677727920,"data":{"finished":1628677817},"status":"terminated"}]
