+++ lib/interface/cli/commands/offline-logs/offload-to-collection.cmd.js
@@ -2,6 +2,7 @@ const { MongoClient, ObjectId } = require("mongodb");
 const moment = require('moment')
 const Command = require('../../Command');
 const cmd = require('./base.cmd');
+const { objectIdFromDate } = require('./utils')
 
 const offloadToCollection = async function(sourceDBObj, collection, targetDB, cutoffDate) {
   const sourceCollectionObj = sourceDBObj.collection(collection)
@@ -32,11 +33,6 @@ const offloadToCollection = async function(sourceDBObj, collection, targetDB, cu
       console.error("The old logs were not deleted from the source collection.")
   }
 }
-
-// converts date to objectId for filter purposes
-const objectIdFromDate = function (date) {
-	return Math.floor(date.getTime() / 1000).toString(16) + "0000000000000000";
-};
 
 const command = new Command({
     command: 'offload-to-collection',
