+++ lib/interface/cli/commands/offline-logs/offload-to-file.cmd.js
@@ -1,9 +1,9 @@
-const { join } = require("path");
-const { writeFileSync } = require("fs");
+
 const { MongoClient, ObjectId } = require("mongodb");
 const moment = require('moment')
 const Command = require('../../Command');
 const cmd = require('./base.cmd');
+const {objectIdFromDate, writeLogsToFile, checkRemnant} = require('./utils')
 const { boolean } = require("yargs");
 
 const offloadToFile = async function(database, collection, chunkDuration, cutoffDate, path) {
@@ -68,25 +68,6 @@ const offloadToFile = async function(database, collection, chunkDuration, cutoff
       });
     }
   }
-}
-
-// converts date to objectId for filter purposes
-const objectIdFromDate = function (date) {
-	return Math.floor(date.getTime() / 1000).toString(16) + "0000000000000000";
-};
-
-const writeLogsToFile = async function(upperBound, lowerBound, collection, logsToArchive, path) {
-  const date2 = moment(upperBound).subtract(1, "days")
-  const fileDateRange = `${lowerBound.format('YYYY-MM-DD')}-${date2.format('YYYY-MM-DD')}`
-  const fileName = `${fileDateRange}-${collection}.json`;
-  const absPath = join(path, fileName);
-  const data = JSON.stringify(logsToArchive);
-  writeFileSync(absPath, data);
-};
-
-const checkRemnant = async function(lowerBound, chunkSize, cutoffDateObj) {
-  const date1 = moment(lowerBound).add( chunkSize, 'days')
-  return Math.ceil(cutoffDateObj.diff(date1 , "days" ));
 }
 
 const command = new Command({
