+++ lib/interface/cli/commands/offline-logs/offload-to-file.spec.js
@@ -1,0 +1,9 @@
+const {objectIdFromDate, writeLogsToFile, checkRemnant} = require('./utils')
+
+descrbie('offload-to-file', () => {
+  describe('writeLogsToFile', () => {
+    it('creates a correct file when blah blah blah', () => {
+      writeLogsToFile()
+    });
+  });
+});
